+++ app/AttributeValue.php
@@ -47,6 +47,24 @@ class AttributeValue extends Model
         'geography_val',
     ];
 
+    const patchRules = [
+        'certainty' => 'integer|between:0,100',
+        'certainty_description' => 'string|nullable'
+    ];
+
+    public function patch($values) {
+        foreach($values as $k => $v) {
+            if($k == 'certainty') {
+                $this->possibility = $v;
+            } else if($k == 'certainty_description') {
+                $this->possibility_description = $v;
+            } else {
+                $this->{$k} = $v;
+            }
+        }
+        $this->save();
+    }
+
     public function context() {
         return $this->belongsTo('App\Context');
     }