+++ resources/views/home.blade.php
@@ -2,6 +2,7 @@
 
 @section('content')
     <main-view
+        :bibliography="{{$bibliography}}"
         :concepts="{{$concepts}}"
         :context-types="{{$contextTypes}}"
         :preferences="{{json_encode($p)}}"